backend/api/requirements.txt
flask>=3.0
httpx>=0.27
python-dotenv>=1.0
openrouter>=0.8
serpapi>=1.0
